+++ src/hooks/useMarkerMap.ts
@@ -18,18 +18,27 @@ type UseMarkerMap = ({
 const useMarkerMap: UseMarkerMap = ({ mapContainer, markerImageSrc, positions, imageSize }) => {
   const { map } = useKakaoMap(mapContainer);
 
-  positions.forEach((position) => {
-    const imgSize = new window.kakao.maps.Size(...imageSize);
-    const markerImage = new window.kakao.maps.MarkerImage(markerImageSrc, imgSize);
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const marker = new window.kakao.maps.Marker({
-      position: position.latlng,
-      map,
-      title: position.name,
-      image: markerImage,
-    });
+  window.kakao.maps.load(() => {
+    if (map) {
+      map.setCenter(new window.kakao.maps.LatLng(37.54, 126.978));
+      map.setLevel(6);
+    }
   });
+
+  if (positions && positions.length) {
+    positions.forEach((position) => {
+      const imgSize = new window.kakao.maps.Size(...imageSize);
+      const markerImage = new window.kakao.maps.MarkerImage(markerImageSrc, imgSize);
+
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const marker = new window.kakao.maps.Marker({
+        position: position.latlng,
+        map,
+        title: position.name,
+        image: markerImage,
+      });
+    });
+  }
 
   return { map };
 };
